+++ NEWS.md
@@ -1,5 +1,7 @@
 # CoRC 0.0.9001
 
+* fix errors with argument 'type' of setCompartments() and setGlobalQuantities()
+
 # CoRC 0.0.1
 
 * Initial public beta release